Update DateTimePicker.kt
Created a date-time picker that is included as a widget in Android, removed the timeList and pmAmList as those will have to be manually populated with list items, greatly increasing technical overhead. This date-time picker will not only be more user friendly (and familiar), but will also allow a user to choose the minute that they either took, or plan to take their medication. Meanwhile, we aren't programming individual minutes into DropDown menus, and the user isn't scrolling forever trying to find the exact time they need.