Update MainActivity.kt
Added a preview for Android Studio in the MainActivity.kt file. This will help when building the main screen in Android Studio, reducing time on the emulator...

This said, is your starting screen a logon button? I would recommend taking a look at the requirements for the project before proceeding much further. I believe we aren't supposed to start with a logon screen?